Moved ReprojectedArraySubset to its own file
